Rename GetAllJPAEntities service
