+++ src/commands/services/mod.rs
@@ -6,6 +6,6 @@ pub mod create_jpa_entity_service;
 pub mod create_jpa_many_to_one_relationship_service;
 pub mod create_jpa_one_to_one_relationship_service;
 pub mod create_jpa_repository_service;
-pub mod get_all_jpa_entities;
+pub mod get_all_jpa_entities_service;
 pub mod get_all_jpa_mapped_superclasses;
 pub mod get_jpa_entity_info_service;
